% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{confluenSequential}
\alias{confluenSequential}
\alias{confluen_WaterGAP3_H}
\alias{confluen_WaterGAP3_N}
\alias{confluen_WaterGAP3_U}
\title{confluen with sequential routing map}
\usage{
confluen_WaterGAP3_H(
  RIVER_water_m3,
  RIVER_length_km,
  RIVER_velocity_km,
  RIVER_inflow_m3,
  CELL_cellNumberStep_int,
  CELL_inflowCellNumberStep_int
)

confluen_WaterGAP3_N(
  RIVER_water_m3,
  RIVER_length_km,
  RIVER_velocity_km,
  RIVER_inflow_m3,
  CELL_cellNumberStep_int,
  CELL_inflowCellNumberStep_int,
  Riverlak_cellNumber_int,
  Riverlak_capacity_m3,
  param_Riverlak_lin_storeFactor
)

confluen_WaterGAP3_U(
  RIVER_water_m3,
  RIVER_length_km,
  RIVER_velocity_km,
  RIVER_inflow_m3,
  CELL_cellNumberStep_int,
  CELL_inflowCellNumberStep_int,
  Riverlak_cellNumber_int,
  Riverlak_capacity_m3,
  Reservoi_cellNumber_int,
  Reservoi_demand_m3,
  Reservoi_capacity_m3,
  Reservoi_meanInflow_m3,
  Reservoi_meanDemand_m3,
  Reservoi_releaseCoefficient_1,
  Reservoi_isIrrigate_01,
  param_Riverlak_lin_storeFactor
)
}
\description{
confluen with sequential routing map
}
