% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dataio}
\alias{dataio}
\alias{save_matbin}
\alias{load_matbin}
\alias{bind_matbin}
\alias{save_vecbin}
\alias{load_vecbin}
\title{Save a matrix (Numeric, Integer, or Logical) to a binary file}
\usage{
save_matbin(matrix, filename)

load_matbin(filename)

bind_matbin(input_files, output_file)

save_vecbin(vector, filename)

load_vecbin(filename)
}
\arguments{
\item{matrix}{The matrix to be saved}

\item{filename}{The binary file to load}

\item{input_files}{Vector of input file paths}

\item{output_file}{Path for the combined output file}

\item{vector}{The vector to be saved}
}
\value{
Either a NumericMatrix, IntegerMatrix, or LogicalMatrix

Either a NumericVector, IntegerVector, or LogicalVector
}
\description{
This function saves a matrix to a binary file with a header indicating
the matrix type (1=numeric, 2=integer, 3=logical), dimensions, and data.
All types are stored using 4 bytes per element.

This function loads a matrix from a binary file that was saved
using save_matbin(). It automatically detects the matrix type.

This function combines several matrix files into one. All matrices
must be of the same type and have matching column counts.

This function saves a vector to a binary file with a header indicating
the vector type (1=numeric, 2=integer, 3=logical) and length.
All types are stored using 4 bytes per element.

This function loads a vector from a binary file that was saved
using save_vecbin(). It automatically detects the vector type.
}
