% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evatransActual}
\alias{evatransActual}
\alias{evatransActual_SupplyRatio}
\alias{evatransActual_SupplyPow}
\alias{evatransActual_VIC}
\alias{evatransActual_GR4J}
\alias{evatransActual_UBC}
\alias{evatransActual_LiangLand}
\alias{evatransActual_LiangSoil}
\alias{evatransActual_WaterGAP3}
\title{**actuall evapotranspiration**}
\usage{
evatransActual_SupplyRatio(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_sur_k
)

evatransActual_SupplyPow(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_sup_k,
  param_evatrans_sup_gamma
)

evatransActual_VIC(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_vic_gamma
)

evatransActual_GR4J(atmos_potentialEvatrans_mm, water_mm, capacity_mm)

evatransActual_UBC(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_ubc_gamma
)

evatransActual_LiangLand(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_lia_gamma
)

evatransActual_LiangSoil(
  atmos_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_evatrans_lia_B
)

evatransActual_WaterGAP3(
  ATMOS_potentialEvatrans_mm,
  water_mm,
  capacity_mm,
  param_EVATRANS_wat_petmax
)
}
\arguments{
\item{water_mm}{(mm/m2/TS) water volume in `soilLy`, `groundLy` or intercept of `landLy`, same as `SOIL_water_mm`, `GROUND_water_mm` or `LAND_interceptWater_mm`}

\item{capacity_mm}{(mm/m2) water storage capacity in `soilLy`, `groundLy` or intercept of `landLy`, same as `SOIL_capacity_mm`, `GROUND_capacity_mm` or`LAND_interceptCapacity_mm`}

\item{param_evatrans_sur_k}{<0.1, 1> parameter for [evatransActual_SupplyRatio()], ratio of potential ET, that is estimated as actually ET}

\item{param_evatrans_sup_k}{<0.1, 1> parameter for [evatransActual_SupplyPow()], ratio of this method}

\item{param_evatrans_sup_gamma}{<1, 5> parameter for [evatransActual_SupplyPow()], exponent of this method}

\item{param_evatrans_vic_gamma}{<0.2, 5> parameter for [evatransActual_VIC()]}

\item{param_evatrans_ubc_gamma}{<0.5, 2> parameter for [evatransActual_UBC()]}

\item{param_evatrans_lia_gamma}{<0.4, 1> parameter for [evatransActual_LiangLand()]}

\item{param_evatrans_lia_B}{<0.01, 3> parameter for [evatransActual_LiangSoil()]}

\item{ATMOS_potentialEvatrans_mm}{(mm/m2/TS) **potential / reference** evapotranspiration}

\item{param_EVATRANS_wat_petmax}{<10, 20> parameter for [evatransActual_WaterGAP()], 10 for arid area, 20 for humid area}
}
\value{
actually ET in (mm/m2/TS)
- evaporation in interception (landLy), `land_evatrans_mm`
- transpiration in root
- evaporation in soil (soilLy), `soil_evatrans_mm`
}
\description{
\loadmathjax
Actual ET, or actual evapotranspiration, is a measure of the amount of water that is lost from the land surface through evaporation and transpiration by plants.

Under the concept of the conceptual HM, the actual ET is always calculated by the potential ET \mjseqn{E_p}, which evaluates the meteorological and landuse (vegetation) situations. 
The second point to consider is the water availability of the land or soil.

So we can give the function from:

\mjsdeqn{E_a = f_{evatransActual}(D_{atms}, D_{lssg})}


to:

\mjsdeqn{E_a = f_{evatransActual}(E_p, W_{lssg}, ...) = k^* E_p}

where
- \mjseqn{E_a} is `land_evatrans_mm` or `soil_evatrans_mm`
- \mjseqn{E_p} is `atmos_potentialEvatrans_mm`
- \mjseqn{k^*} is estimated ratio.

Then the different `evatransActual` methods will estimate the ratio \mjseqn{k^*}.

The output density distribution from 7 methods:
}
\details{
# **_SupplyRatio**: 


The water content (the ratio to the maximal capacity) 
is considered as th main factors for the ratio \mjseqn{k^*}.
\mjsdeqn{k^* = k  \frac{W}{C}}
where
  - \mjseqn{W} is water volume in (mm/m2/TS), `water_mm`, `land_interceptWater_mm`, `soil_water_mm`
  - \mjseqn{C} is water capacity in (mm/m2), `capacity_mm`, `land_interceptCapacity_mm`, `soil_capacity_mm`
  - \mjseqn{k} is `param_evatrans_sur_k`

# **_SupplyPow**: 


The water content (the ratio to the maximal capacity) 
is considered as th main factors for the ratio \mjseqn{k^*}.
\mjsdeqn{k^* = k  \left(\frac{W}{C}\right)^\gamma}
where
  - \mjseqn{k} is `param_evatrans_sup_k`
  - \mjseqn{\gamma} is `param_evatrans_sup_gamma`

# **_VIC** \insertCite{VIC_Wood_1992}{HydroGallery}: 


This method is similar with [evatransActual_SupplyPow()], estimate the water content in the storage.
\mjsdeqn{k^* = 1-\left(1-\frac{W}{C}\right)^{\gamma}}
where
  - \mjseqn{\gamma} is `param_evatrans_vic_gamma`

# **_GR4J** \insertCite{GR4J_Perrin_2003}{HydroGallery}: 


It is a little different than other method, it estimate not the ratio \mjseqn{f},
rather dieectly a equation with potential ET and water content.
And it need **no parameter**.
\mjsdeqn{E_a = \frac{W\left(2-\frac{W}{C}\right)\tanh \left(\frac{E_p}{C}\right)}{1 + \left(1-\frac{W}{C}\right)\tanh \left(\frac{E_p}{C}\right)}}

# **_UBC** \insertCite{UBC_Quick_1977}{HydroGallery}: 


It estimates the water content in the storage. 
(This is a little different than original, the parameter `P0AGEN` is replaced by \mjseqn{\frac{C}{\gamma}}.)
\mjsdeqn{k^* = 10^{\gamma \frac{W-C}{C}}}
where
  - \mjseqn{\gamma} is `param_evatrans_ubc_gamma`

# **_LiangLand** \insertCite{VIC2_Liang_1994}{HydroGallery}: 


It is also a similar method like [evatransActual_SupplyPow()], 
but it will estimate the supply ability agian, whwn the water is still not enough.
\mjsdeqn{E_a^* = \left(\frac{W}{C}\right)^\gamma E_p}
\mjsdeqn{E_a = \min \left(1, \frac{W}{E_a^*}\right) E_a^*}
where
  - \mjseqn{E_l^*} is the first estimated actuall ET
  - \mjseqn{E_l} is actuall ET from land, `land_evatrans_mm`
  - \mjseqn{\gamma} is `param_evatrans_lia_gamma`

# **_LiangSoil** \insertCite{VIC2_Liang_1994}{HydroGallery}: 


It estimates the water content in the storage. 
(This is a little different than original, the parameter `P0AGEN` is replaced by \mjseqn{\frac{C}{\gamma}}.)
\mjsdeqn{k^* = \int_{0}^{A_{s}} {\rm d} A + \int_{A_{s}}^{1} \frac{i_{0}}{i_{m} [1-(1-A)^{1 / B} ]} {\rm d} A }
where
  - \mjseqn{B} is `param_evatrans_lia_B`
  - \mjseqn{A} is fraction of area

![](liang_evatransSoil.png)
}
\references{
\insertAllCited{}
}
