% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evatransPotential}
\alias{evatransPotential}
\alias{landLeafAreaIndex_WaterGAP3}
\alias{evatransPotential_TurcWendling}
\alias{evatransPotential_Linacre}
\alias{evatransPotential_FAO56}
\title{**potential evapotranspiration**}
\usage{
landLeafAreaIndex_WaterGAP3(
  ATMOS_temperature_Cel,
  ATMOS_precipitation_mm,
  CELL_latitude_deg,
  LAND_growUpDay_d,
  LAND_leafAreaIndexMin_,
  LAND_leafAreaIndexMax_,
  Time_dayOfYear_d
)

evatransPotential_TurcWendling(
  atmos_temperature_Cel,
  atmos_solarRadiat_MJ,
  param_evatrans_tur_k
)

evatransPotential_Linacre(
  atmos_temperature_Cel,
  atmos_relativeHumidity_1,
  land_latitude_Degree,
  land_elevation_m,
  land_albedo_1
)

evatransPotential_FAO56(
  atmos_temperature_Cel,
  atmos_vaporPress_hPa,
  atmos_saturatVaporPress_hPa,
  atmos_netRadiat_MJ,
  atmos_windSpeed2m_m_s,
  land_elevation_m
)
}
\arguments{
\item{ATMOS_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{ATMOS_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{LAND_growUpDay_d}{growup day}

\item{LAND_leafAreaIndexMin_, LAND_leafAreaIndexMax_}{minimal and maximal LAI}

\item{param_evatrans_tur_k}{<0.6, 1> parameter for [evatransPotential_TurcWendling()], higher value when closer to the sea}
}
\value{
potential evapotranspiration (mm/m2)
}
\description{
The concept

The concept of potential evapotranspiration (ET) estimates the ability of water lost from the soil and vegetation in an area due to evaporation and transpiration. 
It assumes that there is always enough water in the ET area to meet the demand for evapotranspiration.
However, the characteristics of the ET area, such as whether it is covered with vegetation or bare soil, can affect the amount of evapotranspiration that occurs. 
In order to accurately estimate potential ET, we need to consider these characteristics. 

But we may not always have access to the necessary information or effective methods to do this.
In these cases, we can use a simplified method known as **reference ET**. 
This method defines the ET area using certain fixed characteristics, such as those provided by the [evatransPotential_FAO56()] function. 
In this situation, we need to provide factors to account for the differences between the actual ET area and the reference ET area.
}
\details{
- **_TurcWendling** \insertCite{ET_TurcWendling_1991}{HydroGallery}: consider only the radiation and temperature as the main factors. 
\mjsdeqn{E_p = \frac{(100 R_s + 3.875 t_h k)\cdot(T + 22)}{150 (T + 123)}}
where
  - \mjseqn{E_p} is potential ET, `atmos_potentialEvatrans_mm`
  - \mjseqn{R_s} is solar radiation, `atmos_solarRadiat_MJ`
  - \mjseqn{t_h} is time step in hour, `time_step_h`
  - \mjseqn{T} is average air temperature, `atmos_temperature_Cel`
  - \mjseqn{k} is `param_evatrans_tur_k`

- **_Linacre** \insertCite{ET_Linacre_1977}{HydroGallery}: consider only the temperature as the main factors. 
\mjsdeqn{E_p = \frac{\frac{100(0.75 - \alpha)(T + 0.006 z)}{100 - \phi} + 15(T - T_d)}{80 - T}}
\mjsdeqn{T_d = T - 20 (1-H_R)}
where
  - \mjseqn{\alpha} is albedo, `land_albedo_1`
  - \mjseqn{z} is elevation, `land_elevation_m`
  - \mjseqn{T_d} is dewpoint temperature,
  - \mjseqn{H_R} is relative humidity, `atmos_relativeHumidity_1`

- **_FAO56** \insertCite{ET_FAO56_1998}{HydroGallery}: consider not only radiation and temperature but also other variable like wind speed
as the main factors. 
\mjsdeqn{E_p =\frac{0.408 \Delta\left(R_n - G\right)+\gamma \frac{900}{T+273} {u}_{2}\left({e}_{{s}}-{e}_{{a}}\right)}{\Delta+\gamma\left(1+0.34 {u}_{2}\right)}}
where
  - \mjseqn{\Delta} is slope vapour pressure curve (kPa °C-1)
  - \mjseqn{R_n} is net radiation, `atmos_netRadiat_MJ`
  - \mjseqn{G} is soil heat flux density
  - \mjseqn{u_2} is wind speed at 2 m height, `atmos_windSpeed2m_m_s`
  - \mjseqn{e_s} is saturation vapour pressure, `atmos_saturatVaporPress_hPa`
  - \mjseqn{e_a} is actual vapour pressure, `atmos_vaporPress_hPa`
  - \mjseqn{\gamma} is psychrometric constant
}
\references{
\insertAllCited{}

\insertAllCited{}
}
