% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroGallery-package.R
\docType{_PACKAGE}
\name{HydroGallery-package}
\alias{HydroGallery-package}
\title{A Brief Introduction to HydroGallery}
\description{
An Rcpp-based package designed to provide a versatile and efficient suite of functions for hydrological modeling and analysis.
}
