% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intercep}
\alias{intercep}
\alias{intercep_Full}
\title{**interception** water from land go into the soil.}
\usage{
intercep_Full(
  atmos_precipitation_mm,
  land_interceptWater_mm,
  land_interceptCapacity_mm
)
}
\value{
intercept_water_mm (mm/m2) intercepted water in this timestep
}
\description{
\loadmathjax

In hydrological modeling, interception refers to the process by which water from precipitation is temporarily retained on the surfaces of vegetation, such as leaves and branches, before being returned to the atmosphere through evaporation or drip.

Under the concept of the conceptional HM, the interception will simply be calculated with the maximal interception of the land.
And the interception water will also not go to the land, but will be evaporated.
The maximal Interception of the canopy is maybe difficult to estimate 
but the process is really simple and there is also not so many method to describe it.
}
\details{
# **_Full** : 



consider only the radiation and temperature as the main factors. 
\mjsdeqn{F_{itcp} = C_{icpt} - W_{icpt}}
where
  - \mjseqn{F_{icp}} is `intercept_water_mm`
  - \mjseqn{C_{icpt}} is `land_intercepCapaciy_mm`
  - \mjseqn{W_{icpt}} is `land_intercepWater_mm`
}
