% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lake}
\alias{lake}
\alias{lake_AcceptPow}
\title{**lake outflow**}
\usage{
lake_AcceptPow(
  Lake_water_m3,
  Lake_capacity_m3,
  param_Lake_acp_storeFactor,
  param_Lake_acp_gamma
)
}
\arguments{
\item{Lake_water_m3}{(m3) lake water volume}

\item{Lake_capacity_m3}{(m3) capacity of the lake}

\item{param_Lake_acp_storeFactor}{<uknow> parameter for [lake_AcceptPow()],}

\item{param_Lake_acp_gamma}{<uknow> parameter for [lake_AcceptPow()],}
}
\value{
outflow (m3)
}
\description{
The concept of lake estimates the waterbody outflow for waternet concentation
}
