% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lakevap}
\alias{lakevap}
\alias{lakeevap_Zhao}
\alias{lakeevap_ZhaoTest}
\title{**lake evaporation**}
\usage{
lakeevap_Zhao(
  ATMOS_solarRadiat_MJ,
  ATMOS_temperature_Cel,
  ATMOS_vaporPress_kPa,
  ATMOS_windSpeed2m_m_s,
  LAND_latitude_Degree,
  LAND_elevation_m,
  Lake_temperature_Cel,
  Lake_depth_m,
  Lake_area_km2,
  Lake_fetchLength_m,
  Time_dayOfYear
)

lakeevap_ZhaoTest(
  ATMOS_solarRadiat_MJ,
  ATMOS_temperature_Cel,
  ATMOS_vaporPress_kPa,
  ATMOS_windSpeed2m_m_s,
  LAND_latitude_Degree,
  LAND_elevation_m,
  Lake_temperature_Cel,
  Lake_depth_m,
  Lake_area_km2,
  Lake_fetchLength_m,
  Time_dayOfYear
)
}
\arguments{
\item{ATMOS_solarRadiat_MJ}{(MJ/m2/TS) the solar radiation that actually reaches the earths surface}

\item{ATMOS_temperature_Cel}{(Cel) the average air temperature in the time phase}

\item{ATMOS_windSpeed2m_m_s}{(m/s) wind speed at 2 m above ground surface}

\item{LAND_elevation_m}{(m) average elevation}

\item{Lake_area_km2}{(km2) area of the lake}
}
\value{
evaporation o flake area (mm / day)
}
\description{
The concept of lake estimates the waterbody outflow for waternet concentation
}
