% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lateral}
\alias{lateral}
\alias{lateral_SupplyPow}
\alias{lateral_SupplyRatio}
\alias{lateral_GR4J}
\alias{lateral_GR4Jfix}
\alias{lateral_ThreshPow}
\alias{lateral_Arno}
\title{**lateral flux**}
\usage{
lateral_SupplyPow(
  ground_water_mm,
  ground_capacity_mm,
  param_lateral_sup_k,
  param_lateral_sup_gamma
)

lateral_SupplyRatio(ground_water_mm, param_lateral_sur_k)

lateral_GR4J(ground_water_mm, ground_capacity_mm, ground_potentialLateral_mm)

lateral_GR4Jfix(
  ground_water_mm,
  ground_capacity_mm,
  ground_potentialLateral_mm,
  param_lateral_grf_gamma
)

lateral_ThreshPow(
  ground_water_mm,
  ground_capacity_mm,
  ground_potentialLateral_mm,
  param_lateral_thp_thresh,
  param_lateral_thp_gamma
)

lateral_Arno(
  ground_water_mm,
  ground_capacity_mm,
  ground_potentialLateral_mm,
  param_lateral_arn_thresh,
  param_lateral_arn_k
)
}
\arguments{
\item{param_lateral_sup_k}{<-1, 1> coefficient parameter for [lateral_SupplyPow()]}

\item{param_lateral_sup_gamma}{<0.01, 5> parameters for [lateral_SupplyPow()]}

\item{param_lateral_sur_k}{<-2, 1> coefficient parameter for [lateral_SupplyRatio()]}

\item{param_lateral_grf_gamma}{<0.01, 5> parameter for [lateral_GR4Jfix()]}

\item{param_lateral_thp_thresh}{<0.1, 0.9> coefficient parameter for [lateral_ThreshPow()]}

\item{param_lateral_thp_gamma}{<0.1, 5> exponential parameter for [lateral_ThreshPow()]}

\item{param_lateral_arn_thresh}{<0.1, 0.9> coefficient parameter for [lateral_ThreshPow()]}

\item{param_lateral_arn_k}{<0.1, 1> exponential parameter for [lateral_ThreshPow()]}
}
\value{
lateral_mm (mm/m2)
}
\description{
\loadmathjax

In hydrological modeling, lateral flow refers to the process by which water flows horizontally through the soil or aquifer, rather than vertically.
It is typically represented by a loss term in the water balance equation, so it also named as groundwater exchange (e.g. GR4J \insertCite{GR4J_Perrin_2003}{HydroGallery}).
The flux of lateral exchange is always calculated (only) by the water in the ground layer \mjseqn{W_{grnd}}. 
Unlike other fluxes, the lateral exchange can be positive or negative, 
with positive indicating a supply from other regions and negative indicating distribution to other regions.

This process is so flexible that we must carefully use it, 
because it can easily destroy the waster balance in the research catchment.

\mjsdeqn{F_{ltrl} = f_{lateral}(D_{grnd})}


to:

\mjsdeqn{F_{ltrl} = f_{lateral}(W_{grnd}, C_{grnd}, ...)}


where
- \mjseqn{W_{grnd}} is `ground_water_mm`
- \mjseqn{C_{grnd}} is `ground_capacity_mm`, but not all the methods need the \mjseqn{C_{grnd}}

The output density distribution from 6 methods:
}
\details{
# **_SupplyPow**: 


\mjsdeqn{F_{ltrl} = k \left( \frac{W_{grnd}}{C_{grnd}} \right)^\gamma  W_{grnd}}
where
  - \mjseqn{k} is `param_lateral_sup_k`
  - \mjseqn{\gamma} is `param_lateral_sup_gamma`

# **_SupplyRatio**: 


\mjsdeqn{F_{ltrl} = k * W_{grnd}}
where
  - \mjseqn{k} is `param_lateral_sur_k`

# **_GR4J** \insertCite{GR4J_Perrin_2003}{HydroGallery}: 


\mjsdeqn{F_{ltrl} = M_{ltrl} \left( \frac{W_{grnd}}{C_{grnd}} \right)^{7/2}  }
where
  - \mjseqn{M_{ltrl}} is `ground_potentialLateral_mm`

# **_GR4Jfix** \insertCite{GR4J_Perrin_2003}{HydroGallery} 


based on `_GR4J` use a new parameter to replace the numer 4: 
\mjsdeqn{F_{ltrl} = M_{ltrl} \left( \frac{W_{grnd}}{C_{grnd}} \right)^\gamma  }
where
  - \mjseqn{\gamma} is `param_lateral_grf_gamma`

# **_ThreshPow** 


based on the `_GR4Jfix` and add the one threshold \mjseqn{\phi_b}: 
\mjsdeqn{F_{ltrl} = 0, \quad \frac{W_{grnd}}{C_{grnd}} < \phi_b}
\mjsdeqn{F_{ltrl} = M_{ltrl} \left(\frac{\frac{W_{grnd}}{C_{grnd}} - \phi_b}{1-\phi_b} \right)^\gamma, \quad \frac{W_{grnd}}{C_{grnd}} \geq \phi_b}
where
  - \mjseqn{\phi_b} is `param_lateral_thp_thresh`
  - \mjseqn{\gamma} is `param_lateral_thp_gamma`

# **_Arno** \insertCite{baseflow_Arno_1991,VIC2_Liang_1994}{HydroGallery} 


has also in two cases divided by a threshold water content \mjseqn{\phi_b}: 
\mjsdeqn{F_{ltrl} = k M_{ltrl} \frac{W_{grnd}}{C_{grnd}}, \quad \frac{W_{grnd}}{C_{grnd}} < \phi_b}
\mjsdeqn{F_{ltrl} = k M_{ltrl} \frac{W_{grnd}}{C_{grnd}} + (1-k) M_{ltrl} \left(\frac{W_{grnd} - W_s}{C_{grnd} - W_s} \right)^2, \quad \frac{W_{grnd}}{C_{grnd}} \geq \phi_b}
\mjsdeqn{W_s = k C_{grnd}}
where
  - \mjseqn{\phi_b} is `param_lateral_arn_thresh`
  - \mjseqn{k} is `param_lateral_arn_k`
}
\references{
\insertAllCited{}
}
