% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nc}
\alias{nc}
\alias{read_nc_2d}
\title{Evalute matrics}
\usage{
read_nc_2d(fileName, varName)
}
\arguments{
\item{fileName}{NC-File name.}

\item{varName}{Variable Name.}
}
\value{
2D-Matrix.
}
\description{
Evalute matrics
}
