% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reservoi}
\alias{reservoi}
\alias{reservoiReleas_Hanasaki}
\alias{reservoiReleasCoefficent_Hanasaki}
\title{**Reservoir release**}
\usage{
reservoiReleas_Hanasaki(
  Reservoi_water_m3,
  Reservoi_inflow_m3,
  Reservoi_demand_m3,
  Reservoi_capacity_m3,
  Reservoi_meanInflow_m3,
  Reservoi_meanDemand_m3,
  Reservoi_isIrrigate_01
)

reservoiReleasCoefficent_Hanasaki(
  Reservoi_water_m3,
  Reservoi_capacity_m3,
  Reservoi_releaseCoefficient_1,
  Reservoi_isOperateStart_01,
  param_Reservoi_han_alpha
)
}
\arguments{
\item{Reservoi_water_m3}{(m3) volume of water in the reservoir}

\item{Reservoi_inflow_m3}{(m3/TS) inflow to the reservoir}

\item{Reservoi_demand_m3}{(m3/TS) water demand from the reservoir}

\item{Reservoi_capacity_m3}{(m3) reservoir capacity}

\item{Reservoi_meanInflow_m3}{(m3/TS) annual mean inflow to the reservoir}

\item{Reservoi_meanDemand_m3}{(m3/TS) annual mean water demand from the reservoir}

\item{Reservoi_isIrrigate_01}{(01) indicates if the reservoir is used for irrigation (0 or 1)}

\item{Reservoi_releaseCoefficient_1}{(-) capacity / mena inflow reference to annual mean water release from the reservoir}

\item{Reservoi_isOperateStart_01}{(01) indicates if reservoir operation has started (0 or 1)}

\item{param_Reservoi_han_alpha}{<0,1> 0.85 parameter for [reservoireleasCoefficent_Hanasaki()],}
}
\value{
outflow (m3)

new Reservoi_releaseCoefficient_1
}
\description{
The concept of river estimates the waterbody outflow for waternet concentration
}
