% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{river}
\alias{river}
\alias{riverout_LinearResorvoir}
\alias{riverlakout_LinearResorvoir}
\title{**river outflow**}
\usage{
riverout_LinearResorvoir(
  RIVER_water_m3,
  RIVER_inflow_m3,
  RIVER_velocity_km,
  RIVER_length_km
)

riverlakout_LinearResorvoir(
  Riverlak_water_m3,
  Riverlak_inflow_m3,
  Riverlak_capacity_m3,
  param_Riverlak_lin_storeFactor
)
}
\arguments{
\item{RIVER_water_m3}{(m3) river water volume}

\item{RIVER_inflow_m3}{(m3/TS) river inflow volume}

\item{RIVER_velocity_km}{(km/TS) river velocity}

\item{RIVER_length_km}{(km) river length}

\item{Riverlak_water_m3}{(m3) water volume in the river-lake system}

\item{Riverlak_inflow_m3}{(m3) inflow volume to the river-lake system}

\item{Riverlak_capacity_m3}{(m3) total capacity of the river-lake system}

\item{param_Riverlak_lin_storeFactor}{<unknown> parameter for [riverlak_LinearResorvoir()]}
}
\value{
outflow (m3)
}
\description{
The concept of river estimates the waterbody outflow for waternet concentration
}
