% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infilt}
\alias{infilt}
\alias{infilt_GR4J}
\alias{infilt_UBC}
\alias{infilt_SupplyRatio}
\alias{infilt_AcceptRatio}
\alias{infilt_SupplyPow}
\alias{infilt_AcceptPow}
\alias{infilt_HBV}
\alias{infilt_XAJ}
\alias{infilt_VIC}
\title{**infiltration**}
\usage{
infilt_GR4J(LAND_water_mm, SOIL_water_mm, SOIL_capacity_mm)

infilt_UBC(
  LAND_water_mm,
  LAND_impermeableFrac_1,
  SOIL_water_mm,
  SOIL_capacity_mm,
  param_INFILT_ubc_P0AGEN
)

infilt_SupplyRatio(
  LAND_water_mm,
  SOIL_water_mm,
  SOIL_capacity_mm,
  param_INFILT_sur_k
)

infilt_AcceptRatio(
  LAND_water_mm,
  SOIL_water_mm,
  SOIL_capacity_mm,
  param_INFILT_acr_k
)

infilt_SupplyPow(
  LAND_water_mm,
  SOIL_water_mm,
  SOIL_capacity_mm,
  param_INFILT_sup_k,
  param_INFILT_sup_gamma
)

infilt_AcceptPow(
  LAND_water_mm,
  SOIL_water_mm,
  SOIL_capacity_mm,
  param_INFILT_acp_k,
  param_INFILT_acp_gamma
)

infilt_HBV(
  LAND_water_mm,
  SOIL_water_mm,
  SOIL_capacity_mm,
  param_INFILT_hbv_beta
)

infilt_XAJ(LAND_water_mm, SOIL_water_mm, SOIL_capacity_mm, param_INFILT_xaj_B)

infilt_VIC(LAND_water_mm, SOIL_water_mm, SOIL_capacity_mm, param_INFILT_vic_B)
}
\arguments{
\item{LAND_water_mm}{(mm/m2) **pounded water** volume in `landLy` and there is no limit, different than `LAND_interceptWater_mm`}

\item{SOIL_water_mm}{(mm/m2) water volume in `soilLy`}

\item{SOIL_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{LAND_impermeableFrac_1}{<0, 1> the maximum impermeable fraction when th soil is fully saturated}

\item{param_INFILT_ubc_P0AGEN}{<0.1, 4> coefficient parameter for [infilt_UBC()]}

\item{param_INFILT_sur_k}{<0.01, 1> coefficient parameter for [infilt_SupplyRatio()]}

\item{param_INFILT_acr_k}{<0.01, 1> coefficient parameter for [infilt_AcceptRatio()]}

\item{param_INFILT_sup_k}{<0.01, 1> coefficient parameter for [infilt_SupplyPow()]}

\item{param_INFILT_sup_gamma}{<0, 1> parameters for [infilt_SupplyPow()]}

\item{param_INFILT_acp_k}{<0.01, 1> coefficient parameter for [infilt_AcceptPow()]}

\item{param_INFILT_acp_gamma}{<0.001, 5> parameters for [infilt_AcceptPow()]}

\item{param_INFILT_hbv_beta}{<0.001, 5> parameters for [infilt_HBV()]}

\item{param_INFILT_xaj_B}{<0.01, 3> parameters for [infilt_XAJ()]}

\item{param_INFILT_vic_B}{<0.01, 3> parameters for [infilt_VIC()]}
}
\value{
flux of infiltration from land surface to soil layer

infilt_mm (mm/m2)
}
\description{
\loadmathjax
In hydrological modeling, infiltration refers to the process by which water from precipitation snowmelt or irrigation enters the soil \insertCite{Handbook_Hydrology_1993}{HydroGallery}. 

Under the concept of the conceptual HM, the flux of infiltration is always calculated by the amount of water on the land \mjseqn{W_{land}}, 
which can be precipitation, precipitation after interception, or precipitation with snowmelt, among others. 
The second point to consider is the water acceptability of the soil layer (\mjseqn{C_{soil} - W_{soil}}).

So we can give the function from:

\mjsdeqn{F_{iflt} = f_{infilt}(D_{land}, D_{soil})}


to:

\mjsdeqn{F_{iflt} = f_{infilt}(W_{land}, W_{soil}, C_{soil}, ...)}

some methods will tread the infiltartion as the part of th pounded water so there is also:

\mjsdeqn{F_{iflt} = k^* W_{land}}


where
- \mjseqn{F_{iflt}} is `infilt_mm`
- \mjseqn{W_{land}} is `LAND_water_mm`
- \mjseqn{W_{soil}} is `SOIL_water_mm`
- \mjseqn{C_{soil}} is `SOIL_capacity_mm`
- \mjseqn{k^*} is estimated ratio.

The output density distribution from 9 methods:
}
\details{
# **_GR4J** \insertCite{GR4J_Perrin_2003}{HydroGallery}: 


\mjsdeqn{F_{iflt}=\frac{C_{soil}\left(1-\left(\frac{W_{soil}}{C_{soil}}\right)^{2}\right) \tanh \left(\frac{W_{land}}{C_{soil}}\right)}{1+\frac{W_{soil}}{C_{soil}} \tanh \left(\frac{W_{land}}{C_{soil}}\right)}}

# **_UBC** \insertCite{UBC_Quick_1977}{HydroGallery}: 


estimate the ratio \mjseqn{k^*} as:
\mjsdeqn{k^* = p_{imper} 10^{\frac{W_{soil}-C_{soil}}{p_{AGEN}}}}
where
  - \mjseqn{p_{imper}} is `LAND_impermeableFrac_1`
  - \mjseqn{p_{AGEN}} is `param_INFILT_ubc_P0AGEN`

# **_SupplyRatio**: 


is a very simple method, which estimate only the pounded water:
\mjsdeqn{k^* = k}
where
  - \mjseqn{k} is `param_INFILT_sur_k`

# **_AcceptRatio**: 


\mjsdeqn{F_{iflt} = k (C_{soil} - W_{soil})}
where
  - \mjseqn{k} is `param_INFILT_acr_k`

# **_SupplyPow**: 


is a very simple method, which estimate only the pounded water:
\mjsdeqn{F_{iflt} = kW_{land}^{\gamma}}
where
  - \mjseqn{k} is `param_INFILT_sup_k`
  - \mjseqn{\gamma} is `param_INFILT_sup_gamma`

# **_AcceptPow**: 


\mjsdeqn{F_{iflt} = k \left(\frac{C_{soil} - W_{soil}}{C_{soil}} \right)^{\gamma}}
where
  - \mjseqn{k} is `param_INFILT_acp_k`
  - \mjseqn{\gamma} is `param_INFILT_acp_gamma`

# **_HBV** \insertCite{HBV_Lindstrom_1997}{HydroGallery}: 


estimate the ratio \mjseqn{k^*} as:
\mjsdeqn{k^* = 1-\left(\frac{W_{soil}}{C_{soil}}\right)^{\beta}}
where
  - \mjseqn{\beta} is `param_INFILT_hbv_beta`

# **_XAJ** \insertCite{XAJ_Zhao_1992}{HydroGallery}:


\mjsdeqn{F_{iflt} = MM  \frac{\left( \frac{MM - AU}{MM} \right)^{B+1} - \left( \frac{MM - AU - W_{land}}{MM} \right)^{B+1}}{B+1}}
\mjsdeqn{AU = MM - \left( \frac{(1 - W_{soil})(B+1)}{MM} \right)^{1 / B - 1}  }
\mjsdeqn{MM = C_{soil}(B+1)  }
where
  - \mjseqn{B} is `param_INFILT_xaj_B`

![](xaj_infilt.png)

# **_VIC** \insertCite{VIC_Wood_1992}{HydroGallery}:


\mjsdeqn{F_{infilt} = \int_{i_{0}}^{i_{0}+P} A(i) {\rm d} i}
\mjsdeqn{i = C_{soil}(B+1) \left[ 1 - (1-A)^{1/B} \right]}
where
  - \mjseqn{B} is `param_INFILT_vic_B`
}
\references{
\insertAllCited{}
}
