% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intercep}
\alias{intercep}
\alias{intercep_Full}
\title{**interception** water from land goes into the soil.}
\usage{
intercep_Full(
  ATMOS_precipitation_mm,
  LAND_interceptWater_mm,
  LAND_interceptCapacity_mm
)
}
\arguments{
\item{ATMOS_precipitation_mm}{(mm/m2/TS) precipitaion volum}

\item{LAND_interceptWater_mm}{(mm/m2) initial water volume that can be **intercepted**}

\item{LAND_interceptCapacity_mm}{(mm/m2) average intercept Capacity (maximal storage capacity)}
}
\value{
intercept_water_mm (mm/m2) intercepted water in this timestep
}
\description{
\loadmathjax

In hydrological modeling, interception refers to the process by which water from precipitation is temporarily retained on the surfaces of vegetation, such as leaves and branches, before being returned to the atmosphere through evaporation or drip.

Under the concept of the conceptual HM, the interception will simply be calculated with the maximal interception of the land.
And the interception water will also not go to the land, but will be evaporated.
The maximal Interception of the canopy is maybe difficult to estimate 
but the process is really simple and there are also not so many methods to describe it.
}
\details{
# **_Full** : 

consider only the radiation and temperature as the main factors. 
\mjsdeqn{F_{itcp} = C_{icpt} - W_{icpt}}
where
  - \mjseqn{F_{icp}} is `intercept_water_mm`
  - \mjseqn{C_{icpt}} is `LAND_intercepCapaciy_mm`
  - \mjseqn{W_{icpt}} is `LAND_intercepWater_mm`
}
