% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lakeout}
\alias{lakeout}
\alias{lakeout_SupplyPow}
\title{**lake outflow**}
\usage{
lakeout_SupplyPow(
  Lake_water_m3,
  Lake_capacity_m3,
  param_Lakeout_sup_storeFactor,
  param_Lakeout_sup_gamma
)
}
\arguments{
\item{Lake_water_m3}{(m3) lake water volume}

\item{Lake_capacity_m3}{(m3) capacity of the lake}

\item{param_Lakeout_sup_storeFactor}{<unknown> parameter for [lakeout_SupplyPow()],}

\item{param_Lakeout_sup_gamma}{<unknown> parameter for [lakeout_SupplyPow()],}
}
\value{
outflow (m3)
}
\description{
The concept of lake estimates the waterbody outflow for waternet concentration
}
