% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{percola}
\alias{percola}
\alias{percola_GR4J}
\alias{percola_GR4Jfix}
\alias{percola_MaxPow}
\alias{percola_ThreshPow}
\alias{percola_Arno}
\alias{percola_BevenWood}
\alias{percola_SupplyPow}
\alias{percola_SupplyRatio}
\alias{percola_WaterGAP3}
\title{**percolation**}
\usage{
percola_GR4J(SOIL_water_mm, SOIL_capacity_mm)

percola_GR4Jfix(SOIL_water_mm, SOIL_capacity_mm, param_PERCOLA_grf_k)

percola_MaxPow(
  SOIL_water_mm,
  SOIL_capacity_mm,
  SOIL_potentialPercola_mm,
  param_PERCOLA_map_gamma
)

percola_ThreshPow(
  SOIL_water_mm,
  SOIL_capacity_mm,
  SOIL_potentialPercola_mm,
  param_PERCOLA_thp_thresh,
  param_PERCOLA_thp_gamma
)

percola_Arno(
  SOIL_water_mm,
  SOIL_capacity_mm,
  SOIL_potentialPercola_mm,
  param_PERCOLA_arn_thresh,
  param_PERCOLA_arn_k
)

percola_BevenWood(
  SOIL_water_mm,
  SOIL_capacity_mm,
  SOIL_fieldCapacityPerc_1,
  SOIL_potentialPercola_mm
)

percola_SupplyPow(
  SOIL_water_mm,
  SOIL_capacity_mm,
  param_PERCOLA_sup_k,
  param_PERCOLA_sup_gamma
)

percola_SupplyRatio(SOIL_water_mm, param_PERCOLA_sur_k)

percola_WaterGAP3(
  LAND_water_mm,
  SOIL_potentialPercola_mm,
  param_PERCOLA_wat_01,
  param_PERCOLA_wat_thresh,
  param_PERCOLA_wat_k
)
}
\arguments{
\item{SOIL_water_mm}{(mm/m2) water volume in `soilLy`}

\item{SOIL_capacity_mm}{(mm/m2) average soil Capacity (maximal storage capacity)}

\item{param_PERCOLA_grf_k}{<0.01, 1> coefficient parameter for [percola_GR4Jfix()]}

\item{SOIL_potentialPercola_mm}{<0.01, 7> (mm/m2/TS) **potential** percolation}

\item{param_PERCOLA_map_gamma}{<0.1, 5> exponential parameter for [percola_MaxPow()]}

\item{param_PERCOLA_thp_thresh}{<0.1, 0.9> coefficient parameter for [percola_ThreshPow()]}

\item{param_PERCOLA_thp_gamma}{<0.1, 5> exponential parameter for [percola_ThreshPow()]}

\item{param_PERCOLA_arn_thresh}{<0.1, 0.9> coefficient parameter for [percola_ThreshPow()]}

\item{param_PERCOLA_arn_k}{<0.1, 1> exponential parameter for [percola_ThreshPow()]}

\item{SOIL_fieldCapacityPerc_1}{<0, 1> the relative ratio that the water content can drainage by gravity}

\item{param_PERCOLA_sup_k}{<0.01, 1> coefficient parameter for [percola_SupplyPow()]}

\item{param_PERCOLA_sup_gamma}{<0, 7> parameter for [percola_SupplyPow()]}

\item{param_PERCOLA_sur_k}{<0.01, 1> coefficient parameter for [percola_SupplyRatio()]}

\item{LAND_water_mm}{(mm/m2) **pounded water** volume in `landLy` and there is no limit, different than `LAND_interceptWater_mm`}

\item{param_PERCOLA_wat_01}{<0, 1> 0: percolation, 1: non percolation [percola_WaterGAP3()]}

\item{param_PERCOLA_wat_thresh}{<10, 15> coefficient parameter for [percola_WaterGAP3()]}

\item{param_PERCOLA_wat_k}{<0.1, 1> exponential parameter for [percola_WaterGAP3()]}
}
\value{
percola_mm (mm/m2)
}
\description{
\loadmathjax

In hydrological modeling, percolation refers to the process by which water from the soil moves downward through the pores and cracks in the soil or rock.
This process is physically driven by a moisture gradient, but this is often simplified in conceptual percolation models \insertCite{Raven_Manual_35}{HydroGallery}.
It can be calculated by the water in the soil layer \mjseqn{W_{soil}},
it can also be tread as the part of the \mjseqn{W_{soil}}.

So we can give the function from:

\mjsdeqn{F_{prcl} = f_{percola}(D_{grnd}, D_{soil})}


to:

\mjsdeqn{F_{prcl} = f_{percola}(W_{soil}, C_{soil}, W_{grnd}, C_{grnd}, ...) = k^* W_{soil}}
\mjsdeqn{F_{prcl} \leq W_{soil}}
\mjsdeqn{F_{prcl} \leq C_{grnd} - W_{grnd}}


where
- \mjseqn{F_{prcl}} is `SOIL_percola_mm`
- \mjseqn{W_{soil}} is `water_SOIL_mm`
- \mjseqn{C_{soil}} is `capacity_SOIL_mm`
- \mjseqn{W_{grnd}} is `ground_water_mm`
- \mjseqn{C_{grnd}} is `capacity_water_mm`
- \mjseqn{k^*} is estimated ratio

The output density distribution from 8 methods:
}
\details{
# **_GR4J** \insertCite{GR4J_Perrin_2003}{HydroGallery}: 


\mjsdeqn{k^* = 1 - \left[ 1 + \left(\frac{4}{9} \frac{W_{soil}}{C_{soil}} \right)^4 \right]^{-1/4}}
where
  - \mjseqn{k^*} is estimated ratio

# **_GR4Jfix** \insertCite{GR4J_Perrin_2003}{HydroGallery}: 


\mjsdeqn{k^* = 1 - \left[ 1 + \left(k \frac{W_{soil}}{C_{soil}} \right)^4 \right]^{-1/4}}
where
  - \mjseqn{k} is `param_PERCOLA_grf_k`

# **_MaxPow**: 


\mjsdeqn{F_{prcl} = M_{prcl} \left(\frac{W_{soil}}{C_{soil}} \right)^\gamma}
where
  - \mjseqn{M_{prcl}} is `SOIL_potentialPercola_mm`
  - \mjseqn{\gamma} is `param_baseflow_map_gamma`

# **_ThreshPow** 


based on the `_MaxPow` and add the one threshold \mjseqn{\phi_b}: 
\mjsdeqn{F_{prcl} = 0, \quad \frac{W_{soil}}{C_{soil}} < \phi_b}
\mjsdeqn{F_{prcl} = M_{prcl} \left(\frac{\frac{W_{soil}}{C_{soil}} - \phi_b}{1-\phi_b} \right)^\gamma, \quad \frac{W_{soil}}{C_{soil}} \geq \phi_b}
where
  - \mjseqn{\phi_b} is `param_PERCOLA_thp_thresh`
  - \mjseqn{\gamma} is `param_PERCOLA_thp_gamma`

# **_Arno** \insertCite{baseflow_Arno_1991,VIC2_Liang_1994}{HydroGallery} 


has also in two cases divided by a threshold water content \mjseqn{\phi_b}:
(*This method is actually not the original method, but an analogy with `baseflow_Arno`*) 
\mjsdeqn{F_{prcl} = k M_{prcl} \frac{W_{soil}}{C_{soil}}, \quad \frac{W_{soil}}{C_{soil}} < \phi_b}
\mjsdeqn{F_{prcl} = k M_{prcl} \frac{W_{soil}}{C_{soil}} + (1-k) M_{prcl} \left(\frac{W_{soil} - W_s}{C_{soil} - W_s} \right)^2, \quad \frac{W_{soil}}{C_{soil}} \geq \phi_b}
\mjsdeqn{W_s = k C_{soil}}
where
  - \mjseqn{\phi_b} is `param_PERCOLA_arn_thresh`
  - \mjseqn{k} is `param_PERCOLA_arn_k`

# **_BevenWood** \insertCite{percola_BevenWood_1983,TOPMODEL_Beven_1995}{HydroGallery}: 


\mjsdeqn{k =  \frac{W_{soil}}{C_{soil} - W_{soil}} \quad {\rm and} \quad k \leq 1}
\mjsdeqn{F_{prcl} = k M_{prcl}}
where
  - \mjseqn{k_{fc}} is `SOIL_fieldCapacityPerc_1`
  - \mjseqn{\gamma} is `param_PERCOLA_sup_gamma`

# **_SupplyPow**: 

\mjsdeqn{k^* = k \left(\frac{W_{soil}}{C_{soil}} \right)^\gamma}
where
  - \mjseqn{k} is `param_PERCOLA_sup_k`
  - \mjseqn{\gamma} is `param_PERCOLA_sup_gamma`

# **_SupplyRatio**: 

\mjsdeqn{k^* = k}
where
  - \mjseqn{k} is `param_PERCOLA_sur_k`
}
\references{
\insertAllCited{}
}
